requirements.txt
google_images_download
pillow
psycopg2-binary
aiohttp
python-dotenv